Split node.go into node_sparse.go with Gather/Scatter operations.
